feature/security: File clean-up. Package renaming. Files updates.
