+++ docker-compose.yml
@@ -11,43 +11,45 @@ services:
     ports:
       - "4566:4566"
 
-    postgres:
-      image: postgres:16.2
-      environment:
-        POSTGRES_DB: ${POSTGRES_DB}
-        POSTGRES_USER: ${POSTGRES_USER}
-        POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
-      networks:
-        - keycloak_network
-      healthcheck:
-        test: ["CMD", "pg_isready", "-U", "${POSTGRES_USER}"]
-        interval: 5s
-        timeout: 5s
-        retries: 3
+  postgres:
+    image: postgres:16.2
+    environment:
+      POSTGRES_DB: ${POSTGRES_DB}
+      POSTGRES_USER: ${POSTGRES_USER}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
+    networks:
+      - keycloak_network
+    healthcheck:
+      test: ["CMD", "pg_isready", "-U", "${POSTGRES_USER}"]
+      interval: 5s
+      timeout: 5s
+      retries: 3
+    ports:
+      - "5432:5432"
 
-    keycloak:
-      image: quay.io/keycloak/keycloak:23.0.6
-      command: start
-      ports:
-        - "9090:8080"
-      environment:
-        KC_HOSTNAME: localhost
-        KC_HOSTNAME_PORT: 9090
-        KC_HOSTNAME_STRICT_BACKCHANNEL: false
-        KC_HTTP_ENABLED: true
-        KC_HOSTNAME_STRICT_HTTPS: false
-        KC_HEALTH_ENABLED: true
-        KEYCLOAK_ADMIN: ${ KEYCLOAK_ADMIN }
-        KEYCLOAK_ADMIN_PASSWORD: ${ KEYCLOAK_ADMIN_PASSWORD }
-        KC_DB: postgres
-        KC_DB_URL: jdbc:postgresql://postgres/${POSTGRES_DB}
-        KC_DB_USERNAME: ${POSTGRES_USER}
-        KC_DB_PASSWORD: ${ POSTGRES_PASSWORD }
-      restart: always
-      depends_on:
-        postgres:
-          condition: service_healthy
-      networks:
+  keycloak:
+    image: quay.io/keycloak/keycloak:23.0.6
+    command: start
+    ports:
+      - "9090:8080"
+    environment:
+      KC_HOSTNAME: localhost
+      KC_HOSTNAME_PORT: 9090
+      KC_HOSTNAME_STRICT_BACKCHANNEL: false
+      KC_HTTP_ENABLED: true
+      KC_HOSTNAME_STRICT_HTTPS: false
+      KC_HEALTH_ENABLED: true
+      KEYCLOAK_ADMIN: ${ KEYCLOAK_ADMIN }
+      KEYCLOAK_ADMIN_PASSWORD: ${ KEYCLOAK_ADMIN_PASSWORD }
+      KC_DB: postgres
+      KC_DB_URL: jdbc:postgresql://postgres/${POSTGRES_DB}
+      KC_DB_USERNAME: ${POSTGRES_USER}
+      KC_DB_PASSWORD: ${ POSTGRES_PASSWORD }
+    restart: always
+    depends_on:
+      postgres:
+        condition: service_healthy
+    networks:
         - keycloak_network
 
 networks:
