+++ src/main/java/com/example/localstack/service/MessageListener.java
@@ -1,6 +1,6 @@
 package com.example.localstack.service;
 
-import com.example.localstack.data.Message;
+import com.example.localstack.data.messageEntities.Message;
 
 public interface MessageListener {
     void handle(Message message);
