+++ src/main/java/com/example/localstack/service/MessagePublisher.java
@@ -1,6 +1,6 @@
 package com.example.localstack.service;
 
-import com.example.localstack.data.Message;
+import com.example.localstack.data.messageEntities.Message;
 
 public interface MessagePublisher {
     void publish(String queueName, Message message);
