+++ src/main/java/com/example/localstack/service/s3Services.java
@@ -3,7 +3,7 @@ package com.example.localstack.service;
 import java.io.IOException;
 import java.io.InputStream;
 
-public interface s3Services {
+public interface S3Services {
     void upload(String bucketName, String key, InputStream inputStream);
     InputStream download(String bucketName, String key) throws IOException;
     String downloadAsString(String bucketName, String key) throws IOException;
