--- src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.application.name=localstack
-app.bucket-name: local-test-bucket
-app.queue-name: local-test-queue
